requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1
psycopg2-binary